+++ .github/workflows/ci.yml
@@ -47,5 +47,5 @@ jobs:
     - name: Install dependencies
       run: go mod download
     - name: Run integration tests
-      run: go test -tags=integration ./opkssh/test/integration -timeout=15m -count=1 -parallel=2 -v
+      run: go test -tags=integration ./test/integration -timeout=15m -count=1 -parallel=2 -v
