+++ test/integration/provider/exampleop.Dockerfile
