+++ test/integration/ssh_server/ubuntu.Dockerfile
